+++ tests/conftest.py
@@ -80,7 +80,7 @@ def temp_dir(test_dir) -> Path:
         pytest.param(
             "pyarrow",
             marks=pytest.mark.skipif(
-                pd.__version__ < "2.0.0",
+                pd.__version__ < "4.0.0",
                 reason="`pandas.options.mode.dtype_backend` "
                 "option not available before 2.0.0",
             ),
@@ -89,7 +89,7 @@ def temp_dir(test_dir) -> Path:
 )
 def pd_backend(test_dir, request) -> str:
     """Use to run test with both pandas backends."""
-    if pd.__version__ > "2.0.0":
+    if pd.__version__ > "4.0.0":
         pd.set_option("mode.dtype_backend", request.param)
     return request.param